Utilise some basic libraries to make the implementation "cleaner".
Unsurprisingly, the itertools impl takes a bit longer as it's
storing quite a bit of state, so perf-wise, this is a bad decision.

It does simplify the implementation by a bit, but on the other hand,
you need to know what a cartesian product means or does.

Most programmers worth their weight in salt know it, but for beginners, less so.

Often times, the simplest solution is the best.